DEMRUM-861: move Module related files into new Module group
