+++ SplunkCustomTracking/Sources/SplunkCustomTracking/Module/CustomTrackingConfiguration.swift
